fix 2 failing cdi tests : jsonb
