+++ glassfish-runner/jsonb-platform-tck/pom.xml
@@ -29,7 +29,7 @@
     <properties>
         <arquillian.junit>1.9.1.Final</arquillian.junit>
         <!-- Use JDK17 to run with GF 8.0.0-JDK17-M5 -->
-        <glassfish.container.version>8.0.0-JDK17-M5</glassfish.container.version>
+        <glassfish.container.version>8.0.0-JDK17-M7</glassfish.container.version>
         <glassfish.home>${project.build.directory}/${glassfish.toplevel.dir}</glassfish.home>
         <glassfish.lib.dir>${glassfish.home}/glassfish/lib</glassfish.lib.dir>
         <!-- Use JDK21 to run with GF 8.0.0-M5 -->
